Updating to match the live version

+++ Models/SensorRequest.go
@@ -1,8 +1,8 @@
 package models
 
 type SensorRequest struct {
-	Location    string  `binding:"required"`
-	SensorValue float32 `binding:"required"`
+	Location    string `binding:"required"`
+	SensorValue int    `binding:"required"`
 	//Could make an enum but lazy
 	SensorType string `binding:"required"`
 }
